+++ src/main/java/de/maxhenkel/car/entity/model/GenericCarModel.java
@@ -9,6 +9,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.entity.EntityRendererManager;
 import net.minecraft.util.math.vector.Vector3d;
+import net.minecraft.util.math.vector.Vector3f;
 
 import java.util.List;
 
@@ -32,18 +33,18 @@ public class GenericCarModel extends OBJEntityRenderer<EntityGenericCar> {
         if (text != null && !text.isEmpty()) {
             matrixStack.push();
             RenderSystem.color4f(1F, 1F, 1F, 1F);
-            drawLicensePlate(entity, text, matrixStack, buffer, packedLight);
+            drawLicensePlate(entity, text, matrixStack, buffer, partialTicks, packedLight);
             matrixStack.pop();
         }
 
         matrixStack.pop();
     }
 
-    private void drawLicensePlate(EntityGenericCar car, String txt, MatrixStack matrixStack, IRenderTypeBuffer buffer, int i) {
+    private void drawLicensePlate(EntityGenericCar car, String txt, MatrixStack matrixStack, IRenderTypeBuffer buffer, float partialTicks, int packedLight) {
         matrixStack.push();
         matrixStack.scale(1F, -1F, 1F);
 
-        translateLicensePlate(car, matrixStack);
+        translateLicensePlate(car, matrixStack, partialTicks);
 
         int textWidth = Minecraft.getInstance().fontRenderer.getStringWidth(txt);
         float textScale = 0.01F;
@@ -52,13 +53,14 @@ public class GenericCarModel extends OBJEntityRenderer<EntityGenericCar> {
 
         matrixStack.scale(textScale, textScale, textScale);
 
-        Minecraft.getInstance().fontRenderer.renderString(txt, 0F, 0F, 0xFFFFFF, false, matrixStack.getLast().getMatrix(), buffer, false, 0, i);
+        Minecraft.getInstance().fontRenderer.renderString(txt, 0F, 0F, 0xFFFFFF, false, matrixStack.getLast().getMatrix(), buffer, false, 0, packedLight);
 
         matrixStack.pop();
     }
 
-    protected void translateLicensePlate(EntityGenericCar entity, MatrixStack matrixStack) {
+    protected void translateLicensePlate(EntityGenericCar entity, MatrixStack matrixStack, float partialTicks) {
         Vector3d offset = entity.getLicensePlateOffset();
+        matrixStack.rotate(Vector3f.YP.rotationDegrees(180F - entity.getYaw(partialTicks)));
         matrixStack.translate(offset.x, offset.y, offset.z);
     }