+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.16.4
+- Fixed license plate rendering